+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2024-04-09
+## [Unreleased] - 2024-05-14
 
 **Breaking Changes**: Due to breaking changes, the next release will be a major release (see the Removed section below for details). Timing of that major release will likely be in the Fall of 2023 to coincide with the planned transition to Java 21 upon its release.
 
@@ -32,10 +32,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed potential integer overflow error in ReinsertionDistance
 * Fixed transient fields in Permutation that shouldn't be transient
 * Classes implementing SequenceSampler interface: methods with probability p of sampling an element fixed to always use specified randomness source (previously incorrectly used default randomness source on some calls).
+* Fixed potential finalizer vulnerability in org.cicirello.sequences.distance.EditDistance, detected by SpotBugs analysis. 
 
 ### Dependencies
 * Bump rho-mu from 3.1.0 to 3.2.0
-* Bump org.cicirello:core from 2.5.0 to 2.6.0
+* Bump org.cicirello:core from 2.5.0 to 2.7.0
 
 ### CI/CD
 * Integrated SpotBugs static analysis into build process.